requirements.txt
datetime
pandas
streamlit
snowflake-connector-python